Also log query on page load
