+++ kahuna/public/js/search/query.js
@@ -12,7 +12,7 @@ import template from './query.html';
 import {syntax} from './syntax/syntax';
 import {grStructuredQuery} from './structured-query/structured-query';
 import { v4 } from 'uuid';
-import { sendTelemetryEvent } from '../services/telemetry';
+import { sendTelemetryEvent, sendTelemetryForQuery } from '../services/telemetry';
 import {structureQuery} from './structured-query/syntax';
 import { search } from '.';
 
@@ -62,6 +62,9 @@ query.controller('SearchQueryCtrl', [
     };
 
     ctrl.resetQuery = resetQuery;
+
+    const { nonFree, uploadedByMe } = ctrl.filter;
+    sendTelemetryForQuery(ctrl.filter.query, nonFree, uploadedByMe)
 
     // Note that this correctly uses local datetime and returns
     // midnight for the local user
@@ -184,47 +187,8 @@ query.controller('SearchQueryCtrl', [
           }
           Object.assign(filter, {nonFree: newNonFree, uploadedByMe: false, uploadedBy: undefined});
         }
-        const structuredQuery = structureQuery(ctrl.filter.query);
-        const searchUuid = v4();
         const { nonFree, uploadedByMe } = ctrl.filter;
-        // nonFree is unfortunately either a boolean, stringified boolean, or undefined
-        const freeToUseOnly = (!(nonFree === 'true' || nonFree === true));
-        const uploadedByMeOnly = (uploadedByMe);
-
-        const sendFilterTelemetryEvent = (key, value) => {
-            sendTelemetryEvent('GRID_FILTER', {
-                type: 'filter', 
-                filterType: 'inclusion',
-                key: key, 
-                value: value,
-                searchUuid
-            }, 1)
-        }
-        if (freeToUseOnly) {  
-            sendFilterTelemetryEvent('freeToUseOnly', 'true')
-        }
-        if (uploadedByMeOnly) {  
-            sendFilterTelemetryEvent('uploadedByMeOnly', 'true')
-        }
-
-        structuredQuery.forEach(queryComponent => {
-            // e.g. filter or search:
-            // search > {type: 'text', value: 'my search'}
-            // filter > {type: 'filter', filterType: 'inclusion', key: 'is', value: 'cool'}
-            const { type } = queryComponent;
-            const formattedType = (type) => {
-                if (type === 'text') { return 'GRID_SEARCH'; }
-                if (type === 'filter') { return 'GRID_FILTER'; }
-                return `GRID_${type.toUpperCase()}`;
-            }
-
-            // In case search is empty, as with a search containing only filters
-            sendTelemetryEvent(formattedType(type), {
-                ...queryComponent, 
-                searchUuid: searchUuid,
-            }, 1);
-        });
-
+        sendTelemetryForQuery(ctrl.filter.query, nonFree, uploadedByMe)
         
 
         $state.go('search.results', filter);
