+++ kahuna/public/js/services/telemetry.ts
@@ -1,5 +1,6 @@
 import {UserTelemetryEventSender, IUserTelemetryEvent} from '@guardian/user-telemetry-client';
 import { v4 } from 'uuid';
+import { structureQuery } from '../search/structured-query/syntax';
 
 const getStoredId = (storage: Storage, key: string): string => {
     const maybeId = storage.getItem(key);
@@ -35,5 +36,49 @@ export const sendTelemetryEvent = (type: string, tags?: IUserTelemetryEvent["tag
         type,
         value,
         tags: {...tags, browserUuid: getBrowserId(), sessionUuid: getSessionId()}
+    });
+};
+
+const sendFilterTelemetryEvent = (key: string, value: string, searchUuid: string) => {
+    sendTelemetryEvent('GRID_FILTER', {
+        type: 'filter',
+        filterType: 'inclusion',
+        key: key,
+        value: value,
+        searchUuid: searchUuid
+    }, 1);
+};
+
+export const sendTelemetryForQuery = (query: string, nonFree?: boolean | string, uploadedByMe?: boolean ) => {
+    const structuredQuery = structureQuery(query);
+    const searchUuid = v4();
+    // nonFree is unfortunately either a boolean, stringified boolean, or undefined
+    const freeToUseOnly = (!(nonFree === 'true' || nonFree === true));
+    const uploadedByMeOnly = (uploadedByMe);
+
+    // Only log for true - matching how these filters work in Grid (only applied when true)
+    if (freeToUseOnly) {
+        sendFilterTelemetryEvent('freeToUseOnly', 'true', searchUuid);
+    }
+    if (uploadedByMeOnly) {
+        sendFilterTelemetryEvent('uploadedByMeOnly', 'true', searchUuid);
+    }
+
+    structuredQuery.forEach(queryComponent => {
+        // e.g. filter or search:
+        // search > {type: 'text', value: 'my search'}
+        // filter > {type: 'filter', filterType: 'inclusion', key: 'is', value: 'cool'}
+        const { type } = queryComponent;
+        const formattedType = (type: string) => {
+            if (type === 'text') { return 'GRID_SEARCH'; }
+            if (type === 'filter') { return 'GRID_FILTER'; }
+            return `GRID_${type.toUpperCase()}`;
+        };
+
+        // In case search is empty, as with a search containing only filters
+        sendTelemetryEvent(formattedType(type), {
+            ...queryComponent,
+            searchUuid: searchUuid
+        }, 1);
     });
 };
